==> 00 - Generate Categorical Features.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np

from sklearn import preprocessing

# Bin definitions
quantiles = [0, 0.25, 0.5, 0.75, 1.0]
ageBins   = [0,20,30,40,50,np.inf]
hrsperhrbin = [0,40,50,60,np.inf]

columns = ['index','train','target','agegroup','fnlwgtgroup','workclass','education','maritalstatus','occupation','relationship','race','sex','cptgaingroup','cptlossgroup','hrsgroup','nativecountry']
finalWeightLabels = ['fnlwgt00','fnlwgt01','fnlwgt02','fnlwgt03','fnlwgt04','fnlwgt05','fnlwgt06','fnlwgt07','fnlwgt08','fnlwgt09']
agelabels = ['ageadultunk','ageadult01','ageadult02','ageadult03','ageadult04']
capitalGainLabels = ['cptgain00','cptgain01','cptgain02','cptgain03']
capitalLossLabels = ['cptloss00','cptloss01','cptloss02','cptloss03']
hoursWeekLabels = ['hrsmissing','parttime','fulltime','overtime']

data = pd.read_csv('data/adult_alldata.csv',low_memory=False)
data.drop(['educationnum'],inplace=True,axis=1)

# Convert data to lowercase
data = data.applymap(lambda x: x.lower().strip() if isinstance(x, str) else x)

data.head()

# Age
data['agegroup'] = pd.cut(data.age,bins=ageBins,labels=agelabels,retbins=False)
data.drop(['age'],inplace=True,axis=1)

## Work Class
data.workclass[data.workclass == '?'] = 'wcmissing'

## Final Weight
data['fnlwgtgroup'] = pd.qcut(data.finalweight, q=10, labels=finalWeightLabels)
data.drop(['finalweight'],inplace=True,axis=1)

# Occupation
data.occupation[data.occupation == '?'] = 'occpmissing'

# Bins for Capital Gain
data['cptgaingroup'] = pd.cut(data.capitalgain,4, labels=capitalGainLabels)
data.drop(['capitalgain'],inplace=True,axis=1)

# Bins for Capital Loss
data['cptlossgroup'] = pd.cut(data.capitalloss,4, labels=capitalLossLabels)
data.drop(['capitalloss'],inplace=True,axis=1)

# Bins for Working Hours
data['hrsgroup'] = pd.cut(data.hoursperweek,bins=hrsperhrbin,labels=hoursWeekLabels)
data.drop(['hoursperweek'],inplace=True,axis=1)

# Native Country
data.nativecountry[data.nativecountry == '?'] = 'countrymissing'

# Mapping dictionary for Target feature
mapping = {1: 'yes', 0: 'no'}
data['target'] = data['target'].replace(mapping)

data.head()

data.shape

# Number of features for the training
data.shape[1]-2

data[columns].to_csv('data/train_data_cat.csv',index=False)
