==> 02 - Create Dataset.ipynb <==
# Requires libraries
import numpy as np
import pandas as pd
import pickle as pk

# Dataset columns
datacols = ['target','agegroup','fnlwgtgroup','workclass','education','maritalstatus','occupation','relationship','race','sex','cptgaingroup','cptlossgroup','hrsgroup','nativecountry']

embeddings = np.array(pd.read_csv('data/embeddings_model_w2v_v3.csv', header=None))

with open('data/word_to_index_v3.pkl', 'rb') as file:
    word_to_index = pk.load(file)

data = pd.read_csv('data/train_data_cat.csv')
data = data[datacols]

# Create the dataset for the clustering process
# Each token in ecah line is retrieved and added to have the row embeddings 

dataset = []

for index, row in data.iterrows():
    person = np.zeros((300))
    for i in range(data.shape[1]):
        token = row.iloc[i].lower()
        idx = word_to_index[token]
        person = person + embeddings[idx,:].reshape(300)      
        
    dataset.append(person)   

person = pd.DataFrame(dataset)
del dataset

person.shape

person.to_csv('data/project_dataset.csv',index=False)
