==> 05 - Cluster Analytics.ipynb <==
# Required Libraries
from sklearn.metrics.pairwise import cosine_similarity
import seaborn as sns
import pickle
import matplotlib.pyplot as plt

import pandas as pd
import numpy as np

import plotly.graph_objects as go

seed = 58
np.random.seed(seed)

# Input vector, returns nearest customer(s)
def Cosine_Similarity(centroid,embeddedData,word_to_index,vocab_size,index_to_word,threshold):
              
    similarity_index = []
    similarity_theta = []

    for i in range(embeddedData.shape[0]):
        
        person = embeddedData.iloc[i,:]
        
        theta_sum = np.dot(centroid, person)
        theta_den = np.linalg.norm(centroid) * np.linalg.norm(person)
        theta = theta_sum / theta_den
        
        if theta >= threshold:
            similarity_index.append(i)
            similarity_theta.append(theta)
        
    return similarity_index, similarity_theta

data      = pd.read_csv('data/data4analytics.csv')
centroids = pd.read_csv('data/kmeans_centroids.csv')
embeddedData = pd.read_csv('data/project_dataset.csv')

with open('data/word_to_index_v3.pkl', 'rb') as fp:
    word_to_index = pickle.load(fp)
    
with open('data/index_to_word_v3.pkl', 'rb') as fp:
    index_to_word = pickle.load(fp)    

VOCAB_SIZE = len(index_to_word)

data.shape

embeddedData.head()

centroids.shape

cluster0 = centroids.iloc[0,:]
cluster1 = centroids.iloc[1,:]

index, thetas = Cosine_Similarity(cluster0,embeddedData,word_to_index,VOCAB_SIZE,index_to_word,0.8)

Similarity_cluster0 = []
for idx in index:
    Similarity_cluster0.append(data.iloc[idx,:])

Similarity_cluster0 = pd.DataFrame(Similarity_cluster0)
Similarity_cluster0['theta'] = thetas

Similarity_cluster0 = Similarity_cluster0[Similarity_cluster0.cluster==0]

Similarity_cluster0 = Similarity_cluster0.sort_values(by=['theta'],ascending=False).copy()

Similarity_cluster0.to_csv("data/cluster0.csv",index=False)

index, thetas = Cosine_Similarity(cluster1,embeddedData,word_to_index,VOCAB_SIZE,index_to_word,0.8)

Similarity_cluster1 = []
for i in index:
    Similarity_cluster1.append(data.iloc[i,:])

Similarity_cluster1 = pd.DataFrame(Similarity_cluster1)
Similarity_cluster1['theta'] = thetas

Similarity_cluster1 = Similarity_cluster1[Similarity_cluster1.cluster==1]

Similarity_cluster1 = Similarity_cluster1.sort_values(by=['theta'],ascending=False).copy()

Similarity_cluster1.to_csv("data/cluster1.csv",index=False)
